requirements.txt
Flask
psycopg2-binary
python-dotenv
PyPDF2
pytesseract
pdf2image
google-generativeai
jsonlines 